add user authorization

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-tasks-filter.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,29 +14,29 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) { }
 
-  getTasks(filterDto : GetTaskFilterDto): Promise<Task[]>{
-return this.taskRepository.getTask(filterDto);
+  getTasks(filterDto : GetTaskFilterDto, user: User): Promise<Task[]>{
+return this.taskRepository.getTask(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user:User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: {id, userId: user}});
     if (!found) {
       throw new NotFoundException(`Task with ID:${id} does not exist`);
     }
     return found;
   }
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({id, userId: user.id});
     if(result.affected === 0){
       throw new NotFoundException(`Task with id ${id} not found`)
     }
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
